+++ integration/samples/secondary/feature-a/feature-a.ts
@@ -1,3 +1,4 @@
 import { SHARED_FEATURE } from '@sample/secondary/shared';
 
 export const FEATURE_A = `Yeah, ${SHARED_FEATURE}`;
+export const NAME = '@sample/secondary/feature-a';