+++ src/lib/ng-package/package.transform.ts
@@ -28,6 +28,7 @@ import {
   isPackage,
   fileUrl,
   fileUrlPath,
+  GlobCache,
 } from './nodes';
 import { discoverPackages } from './discover-packages';
 import { createFileWatch } from '../file-system/file-watcher';
@@ -35,6 +36,7 @@ import { NgPackagrOptions } from './options.di';
 import { flatten } from '../utils/array';
 import { copyFile } from '../utils/copy';
 import { ensureUnixPath } from '../utils/path';
+import { FileCache } from '../file-system/file-cache';
 
 /**
  * A transformation for building an npm package:
@@ -94,11 +96,7 @@ export const packageTransformFactory = (
       const ngPkg = graph.get(pkgUri) as PackageNode;
       const entryPoints = [ngPkg.data.primary, ...ngPkg.data.secondaries].map(entryPoint => {
         const { destinationFiles, moduleId } = entryPoint;
-        const node = new EntryPointNode(
-          ngUrl(moduleId),
-          ngPkg.cache.sourcesFileCache,
-          ngPkg.cache.analysisSourcesFileCache,
-        );
+        const node = new EntryPointNode(ngUrl(moduleId), ngPkg.cache.sourcesFileCache);
         node.data = { entryPoint, destinationFiles };
         node.state = 'dirty';
         ngPkg.dependsOn(node);
@@ -130,11 +128,14 @@ const watchTransformFactory = (
       const { data, cache } = graph.find(isPackage) as PackageNode;
       return createFileWatch(data.src, [data.dest]).pipe(
         tap(fileChange => {
-          const { filePath } = fileChange;
-          const { sourcesFileCache, analysisSourcesFileCache } = cache;
+          const { filePath, event } = fileChange;
+          const { sourcesFileCache } = cache;
           const cachedSourceFile = sourcesFileCache.get(filePath);
 
           if (!cachedSourceFile) {
+            if (event === 'unlink' || event === 'add') {
+              cache.globCache = regenerateGlobCache(sourcesFileCache);
+            }
             return;
           }
 
@@ -152,7 +153,6 @@ const watchTransformFactory = (
 
           // delete node that changes
           nodesToClean.forEach(node => {
-            analysisSourcesFileCache.delete(fileUrlPath(node.url));
             sourcesFileCache.delete(fileUrlPath(node.url));
           });
 
@@ -165,6 +165,11 @@ const watchTransformFactory = (
               sourcesFileCache.delete(metadata);
             }
           });
+
+          // Regenerate glob cache
+          if (event === 'unlink' || event === 'add') {
+            cache.globCache = regenerateGlobCache(sourcesFileCache);
+          }
         }),
         debounceTime(200),
         tap(() => log.msg(FileChangeDetected)),
@@ -253,3 +258,15 @@ const scheduleEntryPoints = (epTransform: Transform): Transform =>
       );
     }),
   );
+
+function regenerateGlobCache(sourcesFileCache: FileCache): GlobCache {
+  const cache: GlobCache = {};
+  sourcesFileCache.forEach((value, key) => {
+    // ignore node_modules and file which don't exists as they are not used by globbing in our case
+    if (value.exists && !key.includes('node_modules')) {
+      cache[key] = 'FILE';
+    }
+  });
+
+  return cache;
+}